% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_change_in_ml.R
\name{calc_change_in_ml}
\alias{calc_change_in_ml}
\title{calc_change_in_ml}
\usage{
calc_change_in_ml(snp.object, partition, indexes)
}
\arguments{
\item{snp.object}{A snp.object containing the processed SNP data.}

\item{partition}{An integer vector indicating a partition of the isolates.}

\item{indexes}{Indexes of the isolates to be moved (must come from one cluster.)}
}
\value{
the best cluster to move indexes to.
}
\description{
Calculate the change in the log marginal likelihood after moving index to each possible cluster
}
