% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_log_ml.R
\name{calc_log_ml}
\alias{calc_log_ml}
\title{calc_log_ml}
\usage{
calc_log_ml(snp.object, partition)
}
\arguments{
\item{snp.object}{A snp.object containing the processed SNP data.}

\item{partition}{An integer vector indicating a partition of the isolates.}
}
\value{
The log marginal likelihood of the given partition.
}
\description{
Calculate the log marginal likelihood assuming a Multinomial-Dirichlet distribution
}
