% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_units_2.R
\name{join_units_2}
\alias{join_units_2}
\title{join_units_2}
\usage{
join_units_2(
  snp.object,
  partition,
  threshold = 1e-05,
  n.cores = 1,
  comb.chache = NULL
)
}
\arguments{
\item{snp.object}{A snp.object containing the processed SNP data.}

\item{partition}{An integer vector indicating an initial partition of the isolates.}

\item{threshold}{The increase in marginal log likelihood required to accept a move.}

\item{n.cores}{The number of cores to use.}

\item{comb.chache}{a matrix recording previous marginal llks of combining clusters}
}
\value{
The best partition after combining two clusters as well as
a boolean value indicating whether a move increased the marginal likelihood.
}
\description{
Peform an iteration of the second move in the algorithm. That is combine two clusters
to improve the marginal likelihood.
}
