% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_stirling2.R
\name{log_stirling2}
\alias{log_stirling2}
\title{log_stirling2}
\usage{
log_stirling2(n, k)
}
\arguments{
\item{n}{positive integer greater than zero}

\item{k}{positive integer between 1 and n}
}
\value{
The log of the Stirling number of the second kind.
}
\description{
Calculate the log of the Stirling number of the second kind
Adapted from the function 'Strlng2' in the CryptRndTest package
}
\references{
Bleick, W.W., Wang, P.C.C., Asymptotics of Stirling Numbers of the Second Kind. Proceedings of the American Mathematical Society (1974), 42(2), 575–580.
}
