% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_search_parallel.R
\name{model_search_parallel}
\alias{model_search_parallel}
\title{model_search_parallel}
\usage{
model_search_parallel(
  snp.object,
  partition,
  round.types,
  quiet = FALSE,
  n.extra.rounds = 0,
  n.cores = 1
)
}
\arguments{
\item{snp.object}{A snp.object containing the processed SNP data.}

\item{partition}{An integer vector indicating an initial starting partition.}

\item{round.types}{A vector indicating which series of moves to make.}

\item{quiet}{Whether to suppress progress information (default=FALSE).}

\item{n.extra.rounds}{The number of additional rounds to perform after the default hierBAPS
settings (default=0). If set to Inf it will run until a local optimum is reached
(this might take a long time).}

\item{n.cores}{The number of cores to use.}
}
\value{
an optimised partition and marginal llk
}
\description{
Clusters DNA alignment using independent loci model
}
