% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preproc_alignment.R
\name{preproc_alignment}
\alias{preproc_alignment}
\title{preproc_alignment}
\usage{
preproc_alignment(snp.matrix)
}
\arguments{
\item{snp.matrix}{A matrix containing SNP data. Rows indicate isolates and columns loci.}
}
\value{
an snp.object
}
\description{
Preprocessed the snp matrix for hierBAPS.
}
