% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_lml_logs.R
\name{save_lml_logs}
\alias{save_lml_logs}
\title{save_lml_logs}
\usage{
save_lml_logs(hb.object, file)
}
\arguments{
\item{hb.object}{The resulting object from runnign hierBAPS}

\item{file}{The file you would like to save the log output to.}
}
\description{
Saves the log marginal likelihoods to a text file.
}
\examples{
snp.matrix <- load_fasta(system.file("extdata", "small_seqs.fa", package = "rhierbaps"))
hb.result <- hierBAPS(snp.matrix, max.depth=2, n.pops=20)
save_lml_logs(hb.result,  file.path(tempdir(), "output_file.txt"))

}
