% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_stirling2.R
\name{log_stirling2}
\alias{log_stirling2}
\title{log_stirling2}
\usage{
log_stirling2(n, k)
}
\arguments{
\item{n}{number of objects}

\item{k}{number of partitions}
}
\value{
log of the Stirling number of the second kind
}
\description{
log_stirling2
}
